Fix wrong bitness

+++ test/test_programs/arithmetic_test/compiled_test_arithm.c
@@ -6,7 +6,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-void init(int number, char* name);
+void init(int number, char *name);
 
 bool check_condition(bool cond);
 
@@ -77,8 +77,8 @@ int main(int argc, char **argv) {
     {
         //div-zero quotient should have all bits set
         init(++num, "DivZero (unsigned)");
-        size_t n = 256;
-        size_t m = 0;
+        uint32_t n = 256;
+        uint32_t m = 0;
         assert_equals(0xFFFFFFFFFFFFFFFF, (n / m), &failed_tests);
     }
 
@@ -168,7 +168,7 @@ int main(int argc, char **argv) {
     m_exit(failed_tests);
 }
 
-void init(int number, char* name) {
+void init(int number, char *name) {
     m_write(1, "Test ", 5);
     printi(number);
     m_write(1, "\t(", 2);
